Auto-generate `.d.ts` with global types for auto-imports
This writes an auto-generated `.d.ts` file to disk containing variable declarations in the `global` namespace so that component types for auto-imported components are available in `.astro` files.

NO INTENTION TO RELEASE THIS
While this integration does import stuff in `.astro` files, that is generally not advisable. Prefer importing explicitly as much as possible. The intended use case is for `.md` and `.mdx` content authoring where types are in any case of no use.